Lift `Schematic` into anticipation (shared with all formats)
Move the `Schematic` typeclass to `anticipation` (sibling of `Encodable`), so it
is shared rather than jacinta-local. The Json-specific givens (primitives,
collections, the auto schema derivation, and the low-priority fused `Encodable &
Schematic`) move into `object JsonSchema` — the `Transport` companion — so they
stay in implicit scope. `soundness` now re-exports `anticipation.Schematic`.

(`JsonPointer.encode` inside `object JsonSchema` is now summoned explicitly,
since the fused given puts a generic `.encode` in that object's lexical scope.)
