+++ lib/anticipation/src/codec/anticipation.Schematic.scala
@@ -30,83 +30,15 @@
 ┃                                                                                                  ┃
 ┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┛
                                                                                                   */
-package jacinta
+package anticipation
 
-import scala.annotation.*
-import scala.compiletime.summonInline
-
-import anticipation.*
 import prepositional.*
-import urticose.*
-import vacuous.*
-import wisteria.*
 
-// The third of the three derivations (encoder-only / schema-only / both): the
-// fused `Encodable & Schematic in Json over JsonSchema` for any product or sum.
-// It is deliberately a single intersection given (not a universal `(Encodable,
-// Schematic) => …` combinator), built by deriving the encoder and the schema from
-// the same structure. Kept at low priority (a supertrait of `Schematic`) so a
-// bare `Schematic over JsonSchema` summon never gets upgraded to require an
-// encoder — the combined given is chosen only when the intersection is requested.
-trait SchematicLow:
-  inline given encodableSchematic: [value: Reflection]
-  =>  value is Encodable & Schematic in Json over JsonSchema =
-    new Encodable with Schematic:
-      type Self = value
-      type Form = Json
-      type Transport = JsonSchema
-      def encoded(value: value): Json = summonInline[value is Encodable in Json].encoded(value)
-      def schema(): JsonSchema = JsonSchema.derived[value].schema()
-
-// `Schematic` describes the schema of a value's encoding; `Transport` is the
-// schema's own representation (e.g. `JsonSchema`), so an instance reads `X is
-// Schematic over JsonSchema`. It carries no `Form` of its own: when fused with an
-// encoder as `Encodable & Schematic in Json over JsonSchema`, the wire format
-// (`Form = Json`) comes from `Encodable`.
-object Schematic extends SchematicLow:
-  given byte: Byte is Schematic over JsonSchema = () => JsonSchema.Integer()
-  given short: Short is Schematic over JsonSchema = () => JsonSchema.Integer()
-  given int: Int is Schematic over JsonSchema = () => JsonSchema.Integer()
-  given long: Long is Schematic over JsonSchema = () => JsonSchema.Integer()
-  given float: Float is Schematic over JsonSchema = () => JsonSchema.Number()
-  given double: Double is Schematic over JsonSchema = () => JsonSchema.Number()
-  given text: Text is Schematic over JsonSchema = () => JsonSchema.String()
-  given email: EmailAddress is Schematic over JsonSchema = () => JsonSchema.String()
-  given boolean: Boolean is Schematic over JsonSchema = () => JsonSchema.Boolean()
-
-  given optional: [value: Schematic over JsonSchema]
-  =>  Optional[value] is Schematic over JsonSchema =
-    () =>
-      value.schema() match
-        case entity: JsonSchema.Object  => entity.copy(optional = true)
-        case entity: JsonSchema.Integer => entity.copy(optional = true)
-        case entity: JsonSchema.Number  => entity.copy(optional = true)
-        case entity: JsonSchema.String  => entity.copy(optional = true)
-        case entity: JsonSchema.Array   => entity.copy(optional = true)
-        case entity: JsonSchema.Boolean => entity.copy(optional = true)
-        case entity: JsonSchema.Null    => entity.copy(optional = true)
-
-  given list: [value: Schematic over JsonSchema]
-  =>  List[value] is Schematic over JsonSchema =
-    () => JsonSchema.Array(items = value.schema())
-
-  given set: [value: Schematic over JsonSchema]
-  =>  Set[value] is Schematic over JsonSchema =
-    () => JsonSchema.Array(items = value.schema())
-
-  given map: [key: Encodable in Text, value: Schematic over JsonSchema]
-  =>  Map[key, value] is Schematic over JsonSchema =
-    () => JsonSchema.Object(additionalProperties = true)
-
-  // Auto-derivation of a schema for any product or sum, mirroring the encoder /
-  // decoder auto-givens. Gated on `Reflection` so it never competes with the
-  // primitive givens above.
-  inline given derived: [value: Reflection] => value is Schematic over JsonSchema =
-    JsonSchema.derived
-
-// `Schematic` describes the schema of a value's encoding. `Transport` is the
-// schema's own representation (e.g. `JsonSchema`). The wire format the schema is
-// *for* is not recorded here — in the fused `Encodable & Schematic in Json over
-// JsonSchema` it comes from `Encodable`'s `Form`.
+// The schema of a value's encoding. `Transport` is the schema's own representation
+// (e.g. `JsonSchema` or a TEL `Tels.Type`), so an instance reads `X is Schematic
+// over JsonSchema`. It is a sibling of `Encodable`; the two are commonly fused as
+// `Encodable & Schematic in Json over JsonSchema`, where the wire format (`Form`)
+// comes from `Encodable`. Format-specific given instances live with their format
+// (e.g. in jacinta / stratiform), not here.
 trait Schematic extends Typeclass, Transportive:
   def schema(): Transport
